Add pip-tools requirement include files
These dependencies are taken from the Pipfile, but without the version
pinning.

In order to help generating requirements.txt files from these files, a
simple Makefile is provided: the "update" target generates the .txt
file from the .in file, and the "sync" target installs or uninstalls
dependencies from the current (virtual) environment.

Note that the requirements-dev.txt file is not managed by pip-tools,
since we don't want to pin the dependencies here (pip-sync requires
that dependencies are pinned).